In this branch implemented functionality for clearing all completed tasks
In this commit
- used modules and its import/export functions
- created a new function to clear all tasks that are marked checked(completed). This function is called when some tasks are marked checked and "clear completed tasks" button is clicked.
