+++ src/modules/clearCompletedTasks.js
@@ -1,0 +1,8 @@
+import { displayTasks, todo } from './displaytasks';
+
+const clearCompletedTasks = () => {
+  todo.clearCompleted();
+  displayTasks();
+};
+
+export default clearCompletedTasks;
