+++ src/modules/deltask.js
@@ -1,0 +1,9 @@
+import { displayTasks, todo } from './displaytasks';
+
+const delTask = (index) => {
+  todo.deleteTask(index);
+  todo.saveTask();
+  displayTasks();
+};
+
+export default delTask;
