+++ src/modules/populate.js
@@ -1,0 +1,12 @@
+import { displayTasks, todo } from './displaytasks';
+
+const populate = () => {
+  if (localStorage.getItem('tasks')) {
+    todo.getStoredTasks();
+    displayTasks();
+  } else {
+    displayTasks();
+  }
+};
+
+export default populate;
